pyproject.toml
[project]
name = "nonebot-plugin-charpic"
version = "0.2.0"
description = "Generate picture(or GIF) by characters"
authors = [
    {name = "1umine", email = "a1628420979@163.com"},
]
dependencies = [
    "Pillow>=9.0.0",
    "httpx>=0.27.0",
    "imageio>=2.9.0,<3.0",
    "nonebot2>=2.0,<3.0",
    "nonebot-plugin-alconna>=0.42.4",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
